my_project/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1.1
djangorestframework==3.13.1
filelock==3.8.0
idna==3.3
platformdirs==2.5.2
pycodestyle==2.9.1
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
virtualenv==20.16.4
